Data.String.Interpolate.Parse

Description

YOU SHOULD NOT USE THIS MODULE.

This is exported mainly so tests can introspect on the implementation.

Documentation

data ParseOutput

Each section here is a list of lines.

Content here is defined by the contiguous sequence of lines begining with the first non-blank line and ending with the last non-blank line

Constructors

ParseOutput

Fields

poHeaderWS :: Lines

poContent :: Lines

poFooterWS :: Lines

parseInput :: String -> Either String ParseOutput

Like parseInterpSegments, but for cases where we need to do more complicated transformations on the input. Separates the interpolation input into its content, whitespace header, and whitespace footer.

parseInterpSegments :: String -> Either String Lines

Given the raw input from a quasiquote, parse it into the information we need to output the actual expression.

Returns an error message if parsing fails.

dosToUnix :: String -> String